==> research/04_model_trainer.ipynb <==
import os

%cd C:\Users\harsh\Downloads\ml_project


%pwd


from dataclasses import dataclass
from pathlib import Path

@dataclass(frozen=True)
class DataIngestionConfig:
    root_dir: Path
    train_data_path: Path
    test_data_path: Path
    model_name: str
    alpha: float
    l1_ratio: float
    target_column: str
    

import sys
import os

sys.path.append(os.path.abspath("src"))
from mlProject.constants import *
from mlProject.utils.common import read_yaml, create_directories


from mlProject.entity.config_entity import ModelTrainerConfig

class ConfigurationManager:
    def __init__(
        self,
        config_filepath = CONFIG_FILE_PATH,
        params_filepath = PARAMS_FILE_PATH,
        schema_filepath = SCHEMA_FILE_PATH):

        self.config = read_yaml(config_filepath)
        self.params = read_yaml(params_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])
    
    def get_model_trainer_config(self) -> ModelTrainerConfig:
        config = self.config.model_trainer
        params = self.params.ElasticNet
        schema =  self.schema.TARGET_COLUMN

        create_directories([config.root_dir])

        model_trainer_config = ModelTrainerConfig(
            root_dir=config.root_dir,
            train_data_path = config.train_data_path,
            test_data_path = config.test_data_path,
            model_name = config.model_name,
            alpha = params.alpha,
            l1_ratio = params.l1_ratio,
            target_column = schema.name
            
        )

        return model_trainer_config
    
    from mlProject.entity.config_entity import ModelTrainerConfig

# test initialization
dummy = ModelTrainerConfig(
    root_dir="root",
    train_data_path="train.csv",
    test_data_path="test.csv",
    model_name="model.joblib",
    alpha=0.5,
    l1_ratio=0.3,
    target_column="target"
)
print(dummy)

    




import os
import pandas as pd
from mlProject import logger
from sklearn.linear_model import ElasticNet
import joblib





class ModelTrainer:
    def __init__(self, config: ModelTrainerConfig):
        self.config = config

    
    def train(self):
        train_data = pd.read_csv(self.config.train_data_path)
        test_data = pd.read_csv(self.config.test_data_path)


        train_x = train_data.drop([self.config.target_column], axis=1)
        test_x = test_data.drop([self.config.target_column], axis=1)
        train_y = train_data[[self.config.target_column]]
        test_y = test_data[[self.config.target_column]]


        lr = ElasticNet(alpha=self.config.alpha, l1_ratio=self.config.l1_ratio, random_state=42)
        lr.fit(train_x, train_y) #Creates an ElasticNet model with alpha (regularization strength) and l1_ratio (balance between L1 and L2).

        joblib.dump(lr, os.path.join(self.config.root_dir, self.config.model_name)) #Saves the trained model to a file using joblib, which is a common way to serialize Python objects efficiently.



try:
    config = ConfigurationManager()
    model_trainer_config = config.get_model_trainer_config()
    model_trainer_config = ModelTrainer(config=model_trainer_config)
    model_trainer_config.train()
except Exception as e:
    raise e
